Impl4 of WL2278 - Dynamic port allocation of cluster nodes
When a node restarts it starts over again with fetching the configuration
-  It is not sure that it can use the "old dynamically allocated port number" again.
-  It should however try to reuse the old one, if not possible it should
allocate a new one. One has to be able to distinguish between portnumbers
specified originally in the config, and ones that has been dynamically added
(the latter may be changed if "busy", but the first cannot be changed).

We use negative portnumbers for ports that are ok to change.


ndb/include/mgmapi/mgmapi_debug.h:
  change prototype for:
  ndb_mgm_set_connection_int_parameter
  ndb_mgm_get_connection_int_parameter
  
  Accept/set an int instead of an unsigned
ndb/include/transporter/TransporterRegistry.hpp:
  Use an int (signed) to represent a port number.
  
  Zero means dynamic (but unassigned)
  >0 means static (defined in configuration)
  <0 means dynamic (and should be converted to positive before use)
ndb/src/common/mgmcommon/IPCConfig.cpp:
  Set up the Transporter with a port number of the correct sign.
ndb/src/common/transporter/TransporterRegistry.cpp:
  start_clients_thread:
   - handle negative port numbers
  
  add_transporter_interface:
   - accept signed integer as port number
  
  start_service:
   - If port is dynamic (<0), try to bind to it.
  	if that fails (e.g. some other process has taken that port),
  	then get a new dynamic port number.
ndb/src/mgmapi/mgmapi.cpp:
  Use a signed integer for value in:
  ndb_mgm_set_connection_int_parameter
  ndb_mgm_get_connection_int_parameter
  
  Cast to a Uint32 for storage/retrieval to/from Properties
ndb/src/mgmsrv/MgmtSrvr.cpp:
  getConnectionDbParameter
   - return value as integer (signed)
   - cast as Uint32 for iter.get
ndb/src/mgmsrv/MgmtSrvr.hpp:
  Update prototype of getConnectionDbParameter
  
  int value (not unsigned anymore)
ndb/src/mgmsrv/Services.cpp:
  MgmApiSession::getConnectionParameter
   - value is now signed

+++ ndb/include/transporter/TransporterRegistry.hpp
@@ -232,12 +232,12 @@ public:
   class Transporter_interface {
   public:
     NodeId m_remote_nodeId;
-    unsigned short m_service_port;
+    int m_service_port;
     const char *m_interface;
   };
   Vector<Transporter_interface> m_transporter_interface;
   void add_transporter_interface(NodeId remoteNodeId, const char *interf,
-		  		 unsigned short port);
+		  		 int port);
   Transporter* get_transporter(NodeId nodeId);
   NodeId get_localNodeId() { return localNodeId; };
